New Themes, Fix Register, fix RSS, fix Download API, cache delete, Location Disable & fix badges
We have 2 new themes: Mario Maker: Beta Revived Overworld and Super Mario Bros Night Theme.
For the RSS and download page, it was pointing to the old URL. this fixes it.
The CacheDelete function now automatically deletes outdated settings, instead of us manually having to do it.
For the Location, basically, if your location field is empty, it wont show the field in the post sidebar.
As for the badges, since in the DB, it randomly switches all our URLs to be 0's, I made it so instead of trying to link to the image "0", it'll show up as a normal badge.