So let me tell you the story.
The story of es_ES_lang.php

So one day, there was a coder called NightYoshi370. He installed StyleCI on his repo in order to improve his codebase. However, StyleCI was mad about the way that file was structured, and wants it to be changed. So because the yoshi was bored in school, he decided to manually fix the file, wishing at every moment that he can find a tool to automatically do it for him.

Also, the correct site layout now shows on the profile, and new pages also have Legacy Redirect.